fix: type stdin offset and reattach fromSeq as uint64 (#173)
The reference declares StdinParams.Offset as *uint64 and ReattachParams.FromSeq
as uint64. claustrum used *int / int, inverting the reference at both ends of
the range: a negative value was accepted where the reference rejects it, and
2^64-1 was rejected where the reference treats it as a legal offset.

The negative case is worse than a missing validation. applyStdin computes
skip = applied - offset, so offset:-1 against applied:0 yields skip=1 and the
payload is resliced data[1:] - the first byte of the child's stdin is silently
dropped, and "applied" reports one less than was sent. offset:-N drops N bytes.

Propagate uint64 through the seq/applied lineage: streamFrame.Seq,
managedProc.seq, stdinApplied, applyStdin, reattach, and the FirstSeq/LastSeq/
StdinApplied/Applied result fields. Field ORDER is untouched and every value is
non-negative, so the marshaled frames are unchanged - the existing goldens are
byte-identical.

+++ integration_test.go
@@ -361,7 +361,7 @@ func lastExit(t *testing.T, frames []streamFrame) streamFrame {
 func assertSeqMonotonic(t *testing.T, frames []streamFrame) {
 	t.Helper()
 	for i, f := range frames {
-		if f.Seq != i+1 {
+		if f.Seq != uint64(i+1) {
 			t.Errorf("frame %d has seq %d, want %d (seqs must be 1..N contiguous)", i, f.Seq, i+1)
 		}
 	}
